refactor: fixture 분리
